s3 helpers and user 'model'
