+++ src/handlers/login.js
@@ -5,13 +5,11 @@
 
 import { DynamoDB } from '@aws-sdk/client-dynamodb'
 import { marshall, unmarshall } from '@aws-sdk/util-dynamodb'
-import { S3Client } from '@aws-sdk/client-s3'
 import jwt from 'jsonwebtoken'
 import bcrypt from 'bcryptjs'
+import { readS3Object } from '../s3-helpers.js'
+import { addUser, getUser } from '../user.js'
 import helpers from '../helpers.js'
-
-const ddb = new DynamoDB({ region: process.env.REGION })
-const s3 = new S3Client({ region: process.env.REGION })
 
 function getInput(event) {
   const body = JSON.parse(event.body)
@@ -29,15 +27,10 @@ function getInput(event) {
 }
 
 async function verifyPassword(username, password) {
-  // Query user by username from DynamoDB
-  const params = {
-    TableName: process.env.TABLE_NAME,
-    Key: marshall({ username }),
-  }
-  const { Item } = await ddb.getItem(params)
-  const user = unmarshall(Item)
+  // Query user by username
+  const user = await getUser(username)
 
-  const pwdOk = await bcrypt.compare(password, user.password)
+  const pwdOk = await bcrypt.compare(password, user.hashedPassword)
   if (pwdOk !== true) {
     throw new helpers.BackendError({
       message: `invalid login or password for ${username}`,
@@ -48,17 +41,12 @@ async function verifyPassword(username, password) {
 
 async function issueTokens(username) {
   // Read public and private keys
-  const privateKeyPem = await helpers.readS3File(
-    s3,
+  const privateKeyPem = await readS3Object(
     process.env.BUCKET_NAME,
     process.env.PRIVATE_KEY_NAME
   )
   const jwks = JSON.parse(
-    await helpers.readS3File(
-      s3,
-      process.env.BUCKET_NAME,
-      process.env.JWKS_JSON_NAME
-    )
+    await readS3Object(process.env.BUCKET_NAME, process.env.JWKS_JSON_NAME)
   )
 
   // Issue JWT tokens
